+++ app/jobs/files_job.rb
@@ -5,9 +5,7 @@ class FilesJob < ApplicationJob
   queue_as :default
 
   def perform
-    @blobs = ActiveStorage::Blob.joins(:attachments).merge(
-      ActiveStorage::Attachment.where(record_type: "Folder")
-    ).order(created_at: :desc).limit(10)
+    @blobs = ActiveStorageBlob.latest_files
     ActionCable.server.broadcast(
       "files_channel",
       latest_uploads: AdminController.render(partial: "files/latest_uploads", locals: { blobs: @blobs })